delete a bunch of unnecessary factories
