--- projects/batfish-common-protocol/src/main/java/org/batfish/specifier/DescriptionRegexInterfaceLocationSpecifierFactory.java
@@ -1,22 +1,0 @@
-package org.batfish.specifier;
-
-import com.google.auto.service.AutoService;
-import java.util.regex.Pattern;
-
-/** A {@link LocationSpecifierFactory} for {@link DescriptionRegexInterfaceLocationSpecifier}. */
-@AutoService(LocationSpecifierFactory.class)
-public final class DescriptionRegexInterfaceLocationSpecifierFactory
-    extends CaseInsensitiveRegexLocationSpecifierFactory {
-  public static final String NAME =
-      DescriptionRegexInterfaceLocationSpecifierFactory.class.getSimpleName();
-
-  @Override
-  public String getName() {
-    return NAME;
-  }
-
-  @Override
-  public LocationSpecifier buildLocationSpecifier(Pattern input) {
-    return new DescriptionRegexInterfaceLocationSpecifier(input);
-  }
-}
