--- projects/batfish-common-protocol/src/main/java/org/batfish/specifier/VrfNameRegexInterfaceLinkLocationSpecifierFactory.java
@@ -1,25 +1,0 @@
-package org.batfish.specifier;
-
-import com.google.auto.service.AutoService;
-import java.util.regex.Pattern;
-
-/**
- * A {@link LocationSpecifierFactory} that builds {@link
- * VrfNameRegexInterfaceLinkLocationSpecifier}s.
- */
-@AutoService(LocationSpecifierFactory.class)
-public final class VrfNameRegexInterfaceLinkLocationSpecifierFactory
-    extends CaseInsensitiveRegexLocationSpecifierFactory {
-  public static final String NAME =
-      VrfNameRegexInterfaceLinkLocationSpecifierFactory.class.getSimpleName();
-
-  @Override
-  public String getName() {
-    return NAME;
-  }
-
-  @Override
-  LocationSpecifier buildLocationSpecifier(Pattern pattern) {
-    return new VrfNameRegexInterfaceLinkLocationSpecifier(pattern);
-  }
-}
