+++ projects/batfish-common-protocol/src/test/java/org/batfish/specifier/IpSpaceSpecifierFactoryTest.java
@@ -1,67 +1,26 @@
 package org.batfish.specifier;
 
-import static org.batfish.specifier.IpSpaceSpecifierFactory.load;
-import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.instanceOf;
-import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
 
-import java.util.regex.Pattern;
-import org.batfish.datamodel.IpWildcard;
-import org.batfish.datamodel.IpWildcardSetIpSpace;
-import org.batfish.datamodel.UniverseIpSpace;
-import org.hamcrest.Matchers;
+import java.util.ServiceConfigurationError;
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.ExpectedException;
 
 public class IpSpaceSpecifierFactoryTest {
+  @Rule public final ExpectedException exception = ExpectedException.none();
+
   @Test
   public void testLoad() {
-    IpSpaceSpecifierFactory loaded = load(new InferFromLocationIpSpaceSpecifierFactory().getName());
-    assertThat(loaded, instanceOf(InferFromLocationIpSpaceSpecifierFactory.class));
+    assertThat(
+        IpSpaceSpecifierFactory.load(FlexibleIpSpaceSpecifierFactory.NAME),
+        instanceOf(FlexibleIpSpaceSpecifierFactory.class));
   }
 
   @Test
-  public void testConstantUniverseIpSpaceSpecifier() {
-    assertThat(
-        new ConstantUniverseIpSpaceSpecifierFactory().buildIpSpaceSpecifier(null),
-        equalTo(new ConstantIpSpaceSpecifier(UniverseIpSpace.INSTANCE)));
-  }
-
-  @Test
-  public void testFlexibleLocationIpSpaceSpecifierFactory() {
-    assertThat(
-        load(FlexibleLocationIpSpaceSpecifierFactory.NAME),
-        instanceOf(FlexibleLocationIpSpaceSpecifierFactory.class));
-    IpSpaceSpecifier actual =
-        new FlexibleLocationIpSpaceSpecifierFactory().buildIpSpaceSpecifier("[vrf(foo)]");
-    assertThat(
-        actual,
-        equalTo(
-            new LocationIpSpaceSpecifier(
-                new InterfaceSpecifierInterfaceLocationSpecifier(
-                    new VrfNameRegexInterfaceSpecifier(Pattern.compile("foo"))))));
-  }
-
-  @Test
-  public void testInferFromLocationIpSpaceSpecifierFactory() {
-    assertThat(
-        new InferFromLocationIpSpaceSpecifierFactory().buildIpSpaceSpecifier(null),
-        is(InferFromLocationIpSpaceSpecifier.INSTANCE));
-  }
-
-  @Test
-  public void testConstantWildcardSetIpSpaceSpecifierFactory() {
-    assertThat(
-        load(ConstantWildcardSetIpSpaceSpecifierFactory.NAME),
-        Matchers.instanceOf(ConstantWildcardSetIpSpaceSpecifierFactory.class));
-    assertThat(
-        new ConstantWildcardSetIpSpaceSpecifierFactory()
-            .buildIpSpaceSpecifier("1.2.3.0/24 - 1.2.3.4"),
-        equalTo(
-            new ConstantIpSpaceSpecifier(
-                IpWildcardSetIpSpace.builder()
-                    .including(new IpWildcard("1.2.3.0/24"))
-                    .excluding(new IpWildcard("1.2.3.4"))
-                    .build())));
+  public void testLoadUnknown() {
+    exception.expect(ServiceConfigurationError.class);
+    IpSpaceSpecifierFactory.load("");
   }
 }
