+++ projects/batfish-common-protocol/src/test/java/org/batfish/specifier/NodeSpecifierFactoryTest.java
@@ -1,11 +1,9 @@
 package org.batfish.specifier;
 
-import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.instanceOf;
 import static org.junit.Assert.assertThat;
 
-import java.util.regex.Pattern;
-import org.batfish.common.BatfishException;
+import java.util.ServiceConfigurationError;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
@@ -14,18 +12,15 @@ public class NodeSpecifierFactoryTest {
   @Rule public final ExpectedException exception = ExpectedException.none();
 
   @Test
-  public void testNameRegexNodeSpecifierFactory() {
+  public void testLoad() {
     assertThat(
-        NodeSpecifierFactory.load(NameRegexNodeSpecifierFactory.NAME),
-        instanceOf(NameRegexNodeSpecifierFactory.class));
-    assertThat(
-        NodeSpecifierFactory.load(NameRegexNodeSpecifierFactory.NAME).buildNodeSpecifier("foo"),
-        equalTo(new NameRegexNodeSpecifier(Pattern.compile("foo"))));
+        NodeSpecifierFactory.load(FlexibleNodeSpecifierFactory.NAME),
+        instanceOf(FlexibleNodeSpecifierFactory.class));
   }
 
   @Test
   public void testUnknownNodeSpecifierFactory() {
-    exception.expect(BatfishException.class);
+    exception.expect(ServiceConfigurationError.class);
     NodeSpecifierFactory.load("");
   }
 }
